implement limit in DAO instead of provd (see details)
filtering is done after using limits, which means there's
a chance that there are results that could be accepted by
the filter, but are never returned by provd.

Quick-fix : implement limiting in the DAO. However, this
adds a performance hit because provd returns all of the devices
on every query

Better solution : implement searching in provd. Like that
filtering and limiting can be done at the same time.