add new validation rules based on context type
Our current database schema imposes a few restrictions
when we associate a line with an extension:

A user can only have one internal extension
for his line. The table user_line doesn't support
line <-> extension n/n relations. This should
be fixed once the table user_line is split into 2 tables
(user_line and line_extension)

When associating an incall, the line must already be associated
with a user. This is because incalls in the database cannot
be redirected towards a line. So, we artificially simulate the
association by redirecting the incall towards a user. This
is harder to fix because you would need to refactor a lot of stuff
in the database AND the webi before being able to directly
associate an incall with a line.