fix find_all_callid_between_date
this function now remove calls that started before start_date. The
reason is that when generating calls for a given hour, the call is only
logged if it is complete.

a call starting at 10 and ending at 11 will be generated when start is
11 but will not be deleted if start is 11, generating a new entry at
every call of fill_db with a start at 11. This scenario usually happen
at the end of the day since start is the last out with statistics - 8 hours.