extension takes priorty over caller id number on sccp line
The webi does not make any distinction on a SCCP device between a
caller id displayed on a device VS the caller id displayed when
calling. The SCCP device must always display the "real" number that
can be used for calling, so we make sure that the SCCP line has
the correct number.