hint_dao: handle users with multiple lines (2/2)
Now, all the extensions referencing a same user will share a similar
hint definition (hint "argument"), whatever the user's lines share the
same extension or not or whatever the line's extension are defined in
the same context or not. This should make hints for user with multiple
lines more predictable.