remove outcall_trunk resource
We can do everything with the relationship of sqlalchemy. This dao are
only boilerplate code

The only situation where we need this "layer" is when we need to do
other action when associate/dissociate. But is not our case in
outcall/trunk. If we want to do other action, we should put
this on "service layer" of confd