add schedule dao
Reason to removing nullable=False and server_default from Schedule:
The name is like a label/display_name, so is not mandatory

Reason to use only one relationship with ScheduleTime, instead of 2
(opened and closed):
With 2 relationships, the mode cannot be set automatically when
inserting values.

Reason to use property type, subtype, actionarg1, actionarg2:
confd is mapped to these values

Reason to use property vs hybrid_property:
When the setter/getter is too complicated, then the sqlalchemy Mapper
does not permit to create an object with these property. Anyways, in
this case, we do not really need hybrid_property

Reason to return the complete list of weekdays, monthdays and month if
None:
It is only to be compatible with the webi. The webi can write None
values in these columns, but is not exposed to confd. Anyways, a None
value is the same as all values.

Reason to return self on 'closed_destination' and 'destination' getter:
It will be a lot of work to adapt confd. It's easier to add them