dialaction: fix weird bug with python3 unorderable
reason: The column categoryval of dialaction is already weird (we put
Int in String column ..), se we already have a fix for some others weird
scenarios. These are fixed with IntAsString class. But this doesn't seem to
convert Int As String before some sqlalchemy operation. So Sqlachemy
try to order key before deleting them, but python3 raise an error when
trying to order int() against str()

This not really clean, but our awful DB requires ..