add tenant filters on extensions
the implementation is different from other resource as an extension has to be
in relation to a context that already has a tenant_uuid. Therefore, the
tenant_uuid of a context is propagated to its extensions.

For context that do not exist in the database, the tenant_uuid of the lowest id
entity is used.

I guess that extensions should eventually be modified such that all contexts
linked to an extension exist in the DB.