hint dao: add a new hint for each user that tracks all lines
if a user is in a group we want all of his lines to get called when the call
is distributed to that user.

since app_queue cannot distribute the call to all of the user's line using `&`
a `Local` channel has to be used. To know if the queue member is available or
not we have to use an hint that monitor all of the user's lines.