endpoint_sip: set the username to match the name
the API used to set the username field input into the name column and did not
allow the user to specify a username.

this change allow the new API to allow the user to create an endpoint with a
name and a username while keeping the same behavior for users using the API the
"old" way.