expire rightcall_members relationship
Before removing back_populates, these "session.expire" were not
necessary because it was some dark magic behind, but it was not a good
dark magic ...

If we had a clean database, then it will be probably not necessary, but
with our awkward database. If we do not do "session.expire", then
(outcall|group) id will be not cast to string when using the same object
and the inverse relationship (call_permission.(group|outcall)) will fail

Association call_permission tests show the issue if we remove this
commit.