meeting: set the tz when inserting the meeting date
utcnow() returns the time without the timezone. when inserting the value into a
datetime field with tz the tz is going to be the tz of the stack when
unspecified.  Adding the tz to the now time explicitly removes the ambiguity and
forces postgres to store the right time. Which is going to be the localized time
with the appropriate offset.