user line: keep group/queue member if secondary line deleted
When a user is a member of a group or a queue the user's main line is the
member.
Not any other lines. When deleting a line, only the primary line should trigger a
removal from the group or queue.

Extensions group members and agent queue members do not fall into this fix