change user loader strategy to selectinload
Why:

* joinedload joins table to the query without filtering before,
creating a lot of cartesian products.  In the case of user which
has tons of relationship, this creates an enormous result set,
and which SQLA pre-allocates, but fails to release after query
In some instance, in can make confd start to swap memory and destroy
performance.

Selectinload is more memory efficient and allows filtering when it
joins by using the IN operator (in this case, avoiding cartesian products)
This strategy works best for 1-N or N-M relationship, which users is
abundant of.