hints: change interface to return a dict of context/result
the old version was called with a context as a parameter and returned a list.
The new version returns a dict containing with keys being the context and the value
the list of hints.

This new interface allows wazo-confgend to query all hints once for each type instead
of doing a DB query for each context which is expensive on a multi-tenant stack with
1000s of contexts.